[Refactor] 최신 쿠폰 페이징 조회 리팩토링 - #101
