+++ src/main/java/com/appcenter/marketplace/domain/coupon/controller/CouponController.java
@@ -3,9 +3,11 @@ package com.appcenter.marketplace.domain.coupon.controller;
 import com.appcenter.marketplace.domain.coupon.dto.res.ClosingCouponRes;
 import com.appcenter.marketplace.domain.coupon.dto.res.CouponMemberRes;
 import com.appcenter.marketplace.domain.coupon.dto.res.CouponPageRes;
+import com.appcenter.marketplace.domain.coupon.dto.res.LatestCouponRes;
 import com.appcenter.marketplace.domain.coupon.service.CouponService;
 import com.appcenter.marketplace.global.common.CommonResponse;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -14,7 +16,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.time.LocalDateTime;
+import java.util.List;
+
 import static com.appcenter.marketplace.global.common.StatusCode.COUPON_FOUND;
+import static com.appcenter.marketplace.global.common.StatusCode.MARKET_FOUND;
 
 @Tag(name = "[쿠폰]", description = "[사장님,회원] 매장 별 쿠폰 조회")
 @RestController
@@ -37,12 +43,28 @@ public class CouponController {
                 .body(CommonResponse.from(COUPON_FOUND.getMessage(),couponService.getCouponList(marketId, couponId, size)));
     }
 
+    @Operation(summary = "최신 등록 쿠폰 조회",
+            description = "최신 쿠폰을 등록순으로 정렬 <br>" +
+                    "처음 요청 시, pageSize만 보내면 됩니다. (기본값은 10입니다) <br>"
+    )
+    @GetMapping("/latest")
+    public ResponseEntity<CommonResponse<CouponPageRes<LatestCouponRes>>> getLatestMarketByCoupon(
+            @Parameter(description = "위에 작성한 marketId의 createdAt (e.g. 2024-11-20T00:59:33.469  OR  2024-11-20T00:59:33.469664 )")
+            @RequestParam(required = false, name = "lastCreatedAt") LocalDateTime lastCreatedAt,
+            @Parameter(description = "각 페이지의 마지막 couponId (e.g. 5)")
+            @RequestParam(required = false, name = "lastCouponId") Long couponId,
+            @RequestParam(defaultValue = "10", name = "pageSize") Integer size) {
+        return ResponseEntity
+                .ok(CommonResponse.from(MARKET_FOUND.getMessage(),
+                        couponService.getLatestCouponPage(lastCreatedAt, couponId,size)));
+    }
+
     @Operation(summary = "마감 임박 쿠폰 리스트",
             description = "마감 임박이 다가온 쿠폰을 조회합니다. <br>" +
                     "처음 요청 시엔 pageSize만 필요합니다. 기본값은 10입니다. <br>" +
                     "만약 쿠폰의 마감일자가 같을 시, 최신 등록 매장 순으로 보여지게 됩니다.")
     @GetMapping("/closing")
-    public ResponseEntity<CommonResponse<CouponPageRes<ClosingCouponRes>>> getClosingTopCouponList(
+    public ResponseEntity<CommonResponse<List<ClosingCouponRes>>> getClosingTopCouponList(
             @RequestParam(defaultValue = "10", name="pageSize") Integer size){
         return ResponseEntity.ok(CommonResponse.from(COUPON_FOUND.getMessage(),
                 couponService.getClosingCouponPage(size)));
