+++ src/main/java/com/appcenter/marketplace/domain/coupon/dto/res/LatestCouponRes.java
@@ -1,0 +1,28 @@
+package com.appcenter.marketplace.domain.coupon.dto.res;
+
+import com.querydsl.core.annotations.QueryProjection;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+public class LatestCouponRes {
+    private final Long couponId;
+    private final String couponName;
+    private final Long marketId;
+    private final String marketName;
+    private final String address;
+    private final String thumbnail;
+    private final LocalDateTime CouponCreatedAt;
+
+    @QueryProjection
+    public LatestCouponRes(Long couponId, String couponName, Long marketId, String marketName, String address, String thumbnail, LocalDateTime couponCreatedAt) {
+        this.couponId = couponId;
+        this.couponName = couponName;
+        this.marketId = marketId;
+        this.marketName = marketName;
+        this.address = address;
+        this.thumbnail = thumbnail;
+        this.CouponCreatedAt = couponCreatedAt;
+    }
+}
