+++ src/main/java/com/appcenter/marketplace/domain/market/controller/MarketController.java
@@ -1,7 +1,9 @@
 package com.appcenter.marketplace.domain.market.controller;
 
 
-import com.appcenter.marketplace.domain.market.dto.res.*;
+import com.appcenter.marketplace.domain.market.dto.res.MarketDetailsRes;
+import com.appcenter.marketplace.domain.market.dto.res.MarketPageRes;
+import com.appcenter.marketplace.domain.market.dto.res.MarketRes;
 import com.appcenter.marketplace.domain.market.service.MarketService;
 import com.appcenter.marketplace.global.common.CommonResponse;
 import io.swagger.v3.oas.annotations.Operation;
@@ -12,9 +14,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
-import static com.appcenter.marketplace.global.common.StatusCode.COUPON_FOUND;
 import static com.appcenter.marketplace.global.common.StatusCode.MARKET_FOUND;
 
 @Tag(name = "[매장]", description = "[사장님,회원] 매장 조회")
@@ -127,35 +127,5 @@ public class MarketController {
 //                        ,marketService.getTopFavoriteMarkets(memberId, size)));
 //    }
 
-    @Operation(summary = "최신 등록 쿠폰 TOP 조회",
-            description = "매장별 최신 등록한 쿠폰을 조회합니다. 기본값은 10개 입니다. 이때, TOP 20으로 변동해야 할 시, count를 20으로 넣어주시면 됩니다. <br>" +
-                    "'최신 등록'이란, 수정포함입니다.<br>" +
-                    " 즉, 사장님이 쿠폰 내용을 수정하거나, 숨김/보기 처리를 하게 되면, 최신 등록 집계에 포함이 됩니다.")
-    @GetMapping("/top-latest-coupon")
-    public ResponseEntity<CommonResponse<List<TopLatestCouponRes>>> getLatestTopCouponList(
-            @RequestParam Long memberId,
-            @RequestParam(defaultValue = "10", name = "size") Integer size) {
-        return ResponseEntity.ok(CommonResponse.from(COUPON_FOUND.getMessage(),
-                marketService.getTopLatestCoupons(memberId, size)));
-    }
 
-    @Operation(summary = "최신 등록 쿠폰의 매장 더보기 조회",
-            description = "조회 기준은 다음과 같습니다. <br>" +
-                    "- 각 매장별 최근에 등록된 쿠폰을 시간순으로 정렬 <br>" +
-                    "즉, 가장 최근에 등록된 쿠폰의 매장 순으로 리스트가 조회됩니다. <br> <br>" +
-                    "처음 요청 시, pageSize만 보내면 됩니다. (기본값은 10입니다) <br>" +
-                    "hasNext가 true일 시, 다음 페이지가 존재합니다.<br>"
-    )
-    @GetMapping("/latest-coupon")
-    public ResponseEntity<CommonResponse<MarketPageRes<LatestCouponRes>>> getLatestMarketByCoupon(
-            @RequestParam(required = true) Long memberId,
-            @Parameter(description = "각 페이지의 마지막 couponId (e.g. 5)")
-            @RequestParam(required = false, name = "lastPageIndex") Long marketId,
-            @Parameter(description = "위에 작성한 marketId의 createdAt (e.g. 2024-11-20T00:59:33.469  OR  2024-11-20T00:59:33.469664 )")
-            @RequestParam(required = false, name = "lastCreatedAt") LocalDateTime lastCreatedAt,
-            @RequestParam(defaultValue = "10", name = "pageSize") Integer size) {
-        return ResponseEntity
-                .ok(CommonResponse.from(MARKET_FOUND.getMessage(),
-                        marketService.getLatestCouponPage(memberId, lastCreatedAt, marketId, size)));
-    }
 }
