+++ src/main/java/com/appcenter/marketplace/domain/market/service/impl/MarketServiceImpl.java
@@ -154,24 +154,6 @@ public class MarketServiceImpl implements MarketService {
 //        return marketRepository.findTopFavoriteMarkets(memberId, size);
 //    }
 
-
-    // 최신 등록 쿠폰 TOP 조회
-    @Override
-    public List<TopLatestCouponRes> getTopLatestCoupons(Long memberId, Integer size) {
-        return marketRepository.findTopLatestCoupons(memberId, size);
-    }
-
-    // 최신 등록 쿠폰의 매장 더보기 조회
-    @Override
-    public MarketPageRes<LatestCouponRes> getLatestCouponPage(Long memberId, LocalDateTime lastCreatedAt, Long lastCouponId, Integer size) {
-        List<LatestCouponRes> resDtoList = marketRepository.findLatestCouponList(memberId, lastCreatedAt, lastCouponId, size);
-
-        if (resDtoList.isEmpty())
-            throw new CustomException(MARKET_NOT_EXIST);
-
-        return checkNextPageAndReturn(resDtoList, size);
-    }
-
     private <T> MarketPageRes<T> checkNextPageAndReturn(List<T> marketResDtoList, Integer size) {
         boolean hasNext = false;
